Row Deletion confirmation implemented

+++ src/com/mysql/request/DeleteRow.java
@@ -2,8 +2,8 @@ package com.mysql.request;
 
 import com.gui.Window;
 import com.mysql.connection.MysqlServerConnect;
-import com.templates.DatabaseListModel;
 
+import javax.swing.*;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -14,22 +14,24 @@ public class DeleteRow {
     private static int AGE;
 
     public DeleteRow() {
-        ResultSet newResult;
-        try {
-            stmt = MysqlServerConnect.connection.prepareStatement("DELETE FROM persons WHERE name = ? AND age = ?");
-            stmt.setString(1, NAME);
-            stmt.setInt(2, AGE);
-            int rows = stmt.executeUpdate();
+        int confirm = JOptionPane.showOptionDialog(Window.jFrame,
+                    "Delete?", "Deleting a row", JOptionPane.YES_NO_OPTION,
+                    JOptionPane.QUESTION_MESSAGE, null, null, null);
+        if (confirm == JOptionPane.YES_OPTION) {
+            try {
+                stmt = MysqlServerConnect.connection.prepareStatement("DELETE FROM persons WHERE name = ? AND age = ?");
+                stmt.setString(1, NAME);
+                stmt.setInt(2, AGE);
+                int rows = stmt.executeUpdate();
 
-            newResult = GetDatabase.getBase();
-            DatabaseListModel newdBase_name = new DatabaseListModel(newResult, 1);
-            DatabaseListModel newdBase_age = new DatabaseListModel(newResult, 2);
-            Window.dBase_name = newdBase_name;
-            Window.dBase_age = newdBase_age;
-            Window.datalist.setModel(Window.dBase_name);
+                int idx = Window.dBase_name.indexOf(NAME);
+                Window.dBase_name.remove(idx);
+                Window.dBase_age.remove(idx);
+                Window.datalist.setModel(Window.dBase_name);
 
-        } catch (SQLException ex) {
-            throw new RuntimeException(ex);
+            } catch (SQLException ex) {
+                throw new RuntimeException(ex);
+            }
         }
     }
